+++ src/addin/client-interfaces/addin-client-callbacks.ts
@@ -16,6 +16,11 @@ export interface AddinClientCallbacks {
   buttonClick?: () => void;
 
   /**
+   * Callback raised indicating context was updated by the host.
+   */
+  updateContext?: (context: any) => void;
+
+  /**
    * Callback raised for tile add-ins indicating that the flyout's next button was clicked.
    */
   flyoutNextClick?: () => void;